+++ app/src/test/kotlin/sh/haven/app/agent/McpDragSelectionTest.kt
@@ -131,6 +131,7 @@ class McpDragSelectionTest {
             desktopSessionRegistry = mockk<sh.haven.core.data.desktop.DesktopSessionRegistry>(relaxed = true),
             usbBroker = mockk<sh.haven.core.usb.UsbBroker>(relaxed = true),
             usbIpServer = mockk<sh.haven.core.usb.UsbIpServer>(relaxed = true),
+            usbDriveVmManager = mockk<sh.haven.app.usb.UsbDriveVmManager>(relaxed = true),
             presentationManager = sh.haven.core.data.agent.AgentPresentationManager(),
             havenUiBridge = mockk(relaxed = true),
             standingPolicyRepository = mockk(relaxed = true),